- Update Xen patches to 2.6.29 final and c/s 854. - patches.xen/sfc-external-sram: enable access to Falcon's   external SRAM (bnc#489105). - patches.xen/sfc-sync-headers: sync Solarflare accelerator   headers (bnc#489105). - Update Xen config files.
suse-commit: d5d0a6f940f9edc34816e64acabd5a438dcadc8b

+++ drivers/xen/sfc_netfront/falcon_event.c
@@ -95,7 +95,7 @@ ef_vi_inline int falcon_rx_check_dup(ef_vi* evq, ef_event* ev_out,
 				     const ef_vi_qword* ev)
 {
 	unsigned q_id = QWORD_GET_U(RX_EV_Q_LABEL, *ev);
-	unsigned desc_ptr = QWORD_GET_U(RX_EV_DESC_PTR, *ev);
+	uint16_t desc_ptr = QWORD_GET_U(RX_EV_DESC_PTR, *ev);
 	ef_rx_dup_state_t* rx_dup_state = &evq->evq_state->rx_dup_state[q_id];
 
 	if(likely( desc_ptr != rx_dup_state->rx_last_desc_ptr )) {
